Added code climate
Don't accept this PR, the code climate badge are linked to my repo, it's just
a suggestion.

This was the first repo of serenata-do-amor project that I was looking and I
didn't find a codeclimate badge, looking in other I saw that you guys already
use codeclimate.

If this is useful for this repo too, I can change the codeclimate badge and you
guy add the repo on codeclimate.

Or just add the badge on the master and close the PR

Thanks for the good work.